+++ packages/eslint-config/.eslintrc.js
@@ -1,14 +1,13 @@
 module.exports = {
 	extends: ['./index.js'],
 	overrides: [
-		{
-			files: ['**/*.ts?(x)'],
-
-			// parserOptions: {
-			// 	project: 'tsconfig.json',
-			// 	tsconfigRootDir: __dirname,
-			// },
-		},
+		// {
+		// 	files: ['**/*.ts?(x)'],
+		// 	parserOptions: {
+		// 		project: 'tsconfig.json',
+		// 		tsconfigRootDir: __dirname,
+		// 	},
+		// },
 	],
 	root: true,
 }
