+++ packages/ui/.storybook/decorators/GoogleMaps.tsx
@@ -1,0 +1,11 @@
+import { type StoryFn } from '@storybook/react'
+
+import { GoogleMapsProvider } from '~ui/providers/GoogleMaps'
+
+export const WithGoogleMaps = (Story: StoryFn) => {
+	return (
+		<GoogleMapsProvider>
+			<Story />
+		</GoogleMapsProvider>
+	)
+}
