+++ packages/ui/.storybook/decorators/index.ts
@@ -1,0 +1,10 @@
+// codegen:start {preset: barrel, include: ./*.tsx}
+export * from './GoogleMaps'
+export * from './I18Next'
+export * from './Layouts'
+export * from './Mantine'
+export * from './SearchState'
+export * from './StrictMode'
+export * from './Trpc'
+export * from './WhyDidYouRender'
+// codegen:end
